+++ packages/cross_file/cross_file/lib/src/file_system/file_system_cross_directory.dart
@@ -64,7 +64,6 @@ base class FileSystemXDirectory extends XDirectory {
   ///     params = PlatformFileSystemXDirectoryCreationParams('my/dir/');
   /// }
   ///
-  ///
   /// final dir = FileSystemXDirectory.fromCreationParams(params);
   /// ```
   /// {@endtemplate}