+++ packages/cross_file/cross_file_platform_interface/lib/src/scoped_storage/platform_scoped_storage_cross_file.dart
@@ -67,8 +67,8 @@ base class PlatformScopedStorageXFileCreationParams extends PlatformXFileCreatio
 /// ```
 mixin PlatformScopedStorageXFileExtension implements PlatformXFileExtension {}
 
-/// Interface for a reference to a local data resource within a device's
-/// scoped storage.
+/// Interface for a reference to a data resource within a device's scoped
+/// storage.
 abstract base class PlatformScopedStorageXFile extends PlatformXFile {
   /// Creates a new [PlatformScopedStorageXFile]
   factory PlatformScopedStorageXFile(PlatformScopedStorageXFileCreationParams params) {